google_ADK_youtube shorts/requirements.txt
google-adk
google-generativeai
python-dotenv
